# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LastPiece API
  version: "1.0.0"
  description: |
    Серверная API. Аутентификация: Bearer JWT (accessToken),
    refreshToken хранится в HttpOnly cookie. Базовый URL зависит от PORT
    (по умолчанию 3000).
servers:
  - url: http://localhost:3000
    description: Default PORT=3000 (см. src/shared/config/index.ts)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: BadRequest
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string, nullable: true }
        details:
          type: object
          nullable: true
          additionalProperties: true
...
    NotificationItem:
      type: object
      properties:
        id: { type: string }
        category: { type: string, enum: [system, auth, seller, product] }
        type: { type: string, enum: [info, success, warning, error] }
        title: { type: string }
        message: { type: string, nullable: true }
        data:
          type: object
          nullable: true
          additionalProperties: true
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
        readAt: { type: string, format: date-time, nullable: true }
    NotificationsListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/NotificationItem' }
        nextCursor: { type: string, nullable: true }
    NotificationsReadRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items: { type: string, minLength: 8 }
    NotificationsUpdatedResponse:
      type: object
      properties:
        updated: { type: integer }
...
paths:
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: Список уведомлений
      parameters:
        - name: unread
          in: query
          schema: { type: integer, description: '1 чтобы получить только непрочитанные' }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: cursor
          in: query
          schema:
            type: string
            description: base64 JSON {"id":"<ObjectId>"}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationsListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/notifications/read:
    post:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: Пометить выбранные уведомления как прочитанные
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationsReadRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationsUpdatedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/notifications/read-all:
    post:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: Пометить все уведомления как прочитанные
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationsUpdatedResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/notifications/{id}:
    delete:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: Удалить уведомление
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Удалено }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ------------- UPLOADS --------------
...